More Work on adding Multiple Pieces
This includes a new turn button & script. Which will eventually be used
to switch to the AI.
